Add asr adapter scripts
